WIP: introduce map selection compilation unit
