+++ src/OpenLoco/CMakeLists.txt
@@ -162,6 +162,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapGenerator/MapGenerator.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapGenerator/ModernTerrainGenerator.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapGenerator/OriginalTerrainGenerator.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapSelection.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/RoadElement.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/SurfaceData.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/SurfaceElement.cpp"
@@ -470,6 +471,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapGenerator/MapGenerator.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapGenerator/ModernTerrainGenerator.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapGenerator/OriginalTerrainGenerator.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/MapSelection.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/QuarterTile.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/RoadElement.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Map/SignalElement.h"
