+++ src/OpenLoco/src/Map/MapSelection.cpp
@@ -1,0 +1,187 @@
+#include "MapSelection.h"
+
+namespace OpenLoco::World
+{
+    static loco_global<int16_t, 0x0050A000> _adjustToolSize;
+
+    static loco_global<coord_t, 0x00F24486> _mapSelectionAX;
+    static loco_global<coord_t, 0x00F24488> _mapSelectionBX;
+    static loco_global<coord_t, 0x00F2448A> _mapSelectionAY;
+    static loco_global<coord_t, 0x00F2448C> _mapSelectionBY;
+    static loco_global<uint16_t, 0x00F2448E> _word_F2448E;
+
+    constexpr uint16_t kMapSelectedTilesSize = 300;
+    static loco_global<Pos2[kMapSelectedTilesSize], 0x00F24490> _mapSelectedTiles;
+
+    // TODO: Return std::optional
+    uint16_t setMapSelectionTiles(const World::Pos2& loc, const uint8_t selectionType)
+    {
+        uint16_t xPos = loc.x;
+        uint16_t yPos = loc.y;
+        uint8_t count = 0;
+
+        if (!Input::hasMapSelectionFlag(Input::MapSelectionFlags::enable))
+        {
+            Input::setMapSelectionFlags(Input::MapSelectionFlags::enable);
+            count++;
+        }
+
+        if (_word_F2448E != selectionType)
+        {
+            _word_F2448E = selectionType;
+            count++;
+        }
+
+        uint16_t toolSizeA = _adjustToolSize;
+
+        if (!toolSizeA)
+            toolSizeA = 1;
+
+        toolSizeA = toolSizeA << 5;
+        uint16_t toolSizeB = toolSizeA;
+        toolSizeB -= 32;
+        toolSizeA = toolSizeA >> 1;
+        toolSizeA -= 16;
+        xPos -= toolSizeA;
+        yPos -= toolSizeA;
+        xPos &= 0xFFE0;
+        yPos &= 0xFFE0;
+
+        if (xPos != _mapSelectionAX)
+        {
+            _mapSelectionAX = xPos;
+            count++;
+        }
+
+        if (yPos != _mapSelectionAY)
+        {
+            _mapSelectionAY = yPos;
+            count++;
+        }
+
+        xPos += toolSizeB;
+        yPos += toolSizeB;
+
+        if (xPos != _mapSelectionBX)
+        {
+            _mapSelectionBX = xPos;
+            count++;
+        }
+
+        if (yPos != _mapSelectionBY)
+        {
+            _mapSelectionBY = yPos;
+            count++;
+        }
+
+        mapInvalidateSelectionRect();
+
+        return count;
+    }
+
+    uint16_t setMapSelectionSingleTile(const World::Pos2& loc, bool setQuadrant)
+    {
+        uint16_t xPos = loc.x & 0xFFE0;
+        uint16_t yPos = loc.y & 0xFFE0;
+        uint16_t cursorQuadrant = Ui::ViewportInteraction::getQuadrantOrCentreFromPos(loc);
+
+        auto count = 0;
+        if (!Input::hasMapSelectionFlag(Input::MapSelectionFlags::enable))
+        {
+            Input::setMapSelectionFlags(Input::MapSelectionFlags::enable);
+            count++;
+        }
+
+        if (setQuadrant && _word_F2448E != cursorQuadrant)
+        {
+            _word_F2448E = cursorQuadrant;
+            count++;
+        }
+        else if (!setQuadrant && _word_F2448E != 4)
+        {
+            _word_F2448E = 4;
+            count++;
+        }
+
+        if (xPos != _mapSelectionAX)
+        {
+            _mapSelectionAX = xPos;
+            count++;
+        }
+
+        if (yPos != _mapSelectionAY)
+        {
+            _mapSelectionAY = yPos;
+            count++;
+        }
+
+        if (xPos != _mapSelectionBX)
+        {
+            _mapSelectionBX = xPos;
+            count++;
+        }
+
+        if (yPos != _mapSelectionBY)
+        {
+            _mapSelectionBY = yPos;
+            count++;
+        }
+
+        mapInvalidateSelectionRect();
+
+        return count;
+    }
+
+    // 0x004610F2
+    void mapInvalidateSelectionRect()
+    {
+        if (Input::hasMapSelectionFlag(Input::MapSelectionFlags::enable))
+        {
+            for (coord_t x = _mapSelectionAX; x <= _mapSelectionBX; x += 32)
+            {
+                for (coord_t y = _mapSelectionAY; y <= _mapSelectionBY; y += 32)
+                {
+                    mapInvalidateTileFull({ x, y });
+                }
+            }
+        }
+    }
+
+    // 0x0046112C
+    void mapInvalidateMapSelectionTiles()
+    {
+        if (!Input::hasMapSelectionFlag(Input::MapSelectionFlags::enableConstruct))
+            return;
+
+        for (uint16_t index = 0; index < kMapSelectedTilesSize; ++index)
+        {
+            auto& position = _mapSelectedTiles[index];
+            if (position.x == -1)
+                break;
+            mapInvalidateTileFull(position);
+        }
+    }
+
+    void setMapSelectionArea(const Pos2& locA, const Pos2& locB)
+    {
+        _mapSelectionAX = locA.x;
+        _mapSelectionAY = locA.y;
+        _mapSelectionBX = locB.x;
+        _mapSelectionBY = locB.y;
+    }
+
+    std::pair<Pos2, Pos2> getMapSelectionArea()
+    {
+        return std::make_pair(Pos2{ _mapSelectionAX, _mapSelectionAY }, Pos2{ _mapSelectionBX, _mapSelectionBY });
+    }
+
+    void setMapSelectionCorner(const MapSelectionType corner)
+    {
+        _word_F2448E = corner;
+    }
+
+    MapSelectionType getMapSelectionCorner()
+    {
+        return _word_F2448E;
+    }
+}
