+++ src/OpenLoco/src/Map/TileManager.cpp
@@ -52,15 +52,6 @@ namespace OpenLoco::World::TileManager
     static loco_global<TileElement*, 0x00F00134> _elementsEnd;
     static loco_global<const TileElement*, 0x00F00158> _F00158;
     static loco_global<uint32_t, 0x00F00168> _F00168;
-    static loco_global<coord_t, 0x00F24486> _mapSelectionAX;
-    static loco_global<coord_t, 0x00F24488> _mapSelectionBX;
-    static loco_global<coord_t, 0x00F2448A> _mapSelectionAY;
-    static loco_global<coord_t, 0x00F2448C> _mapSelectionBY;
-    static loco_global<uint16_t, 0x00F2448E> _word_F2448E;
-    static loco_global<int16_t, 0x0050A000> _adjustToolSize;
-
-    constexpr uint16_t mapSelectedTilesSize = 300;
-    static loco_global<Pos2[mapSelectedTilesSize], 0x00F24490> _mapSelectedTiles;
 
     // 0x0046902E
     void removeSurfaceIndustry(const Pos2& pos)
@@ -110,29 +101,6 @@ namespace OpenLoco::World::TileManager
     std::span<TileElement> getElements()
     {
         return std::span<TileElement>(static_cast<TileElement*>(_elements), getElementsEnd());
-    }
-
-    void setMapSelectionArea(const Pos2& locA, const Pos2& locB)
-    {
-        _mapSelectionAX = locA.x;
-        _mapSelectionAY = locA.y;
-        _mapSelectionBX = locB.x;
-        _mapSelectionBY = locB.y;
-    }
-
-    std::pair<Pos2, Pos2> getMapSelectionArea()
-    {
-        return std::make_pair(Pos2{ _mapSelectionAX, _mapSelectionAY }, Pos2{ _mapSelectionBX, _mapSelectionBY });
-    }
-
-    void setMapSelectionCorner(const uint8_t corner)
-    {
-        _word_F2448E = corner;
-    }
-
-    uint8_t getMapSelectionCorner()
-    {
-        return _word_F2448E;
     }
 
     TileElement* getElementsEnd()
@@ -644,161 +612,12 @@ namespace OpenLoco::World::TileManager
         return owner;
     }
 
-    // TODO: Return std::optional
-    uint16_t setMapSelectionTiles(const World::Pos2& loc, const uint8_t selectionType)
-    {
-        uint16_t xPos = loc.x;
-        uint16_t yPos = loc.y;
-        uint8_t count = 0;
-
-        if (!Input::hasMapSelectionFlag(Input::MapSelectionFlags::enable))
-        {
-            Input::setMapSelectionFlags(Input::MapSelectionFlags::enable);
-            count++;
-        }
-
-        if (_word_F2448E != selectionType)
-        {
-            _word_F2448E = selectionType;
-            count++;
-        }
-
-        uint16_t toolSizeA = _adjustToolSize;
-
-        if (!toolSizeA)
-            toolSizeA = 1;
-
-        toolSizeA = toolSizeA << 5;
-        uint16_t toolSizeB = toolSizeA;
-        toolSizeB -= 32;
-        toolSizeA = toolSizeA >> 1;
-        toolSizeA -= 16;
-        xPos -= toolSizeA;
-        yPos -= toolSizeA;
-        xPos &= 0xFFE0;
-        yPos &= 0xFFE0;
-
-        if (xPos != _mapSelectionAX)
-        {
-            _mapSelectionAX = xPos;
-            count++;
-        }
-
-        if (yPos != _mapSelectionAY)
-        {
-            _mapSelectionAY = yPos;
-            count++;
-        }
-
-        xPos += toolSizeB;
-        yPos += toolSizeB;
-
-        if (xPos != _mapSelectionBX)
-        {
-            _mapSelectionBX = xPos;
-            count++;
-        }
-
-        if (yPos != _mapSelectionBY)
-        {
-            _mapSelectionBY = yPos;
-            count++;
-        }
-
-        mapInvalidateSelectionRect();
-
-        return count;
-    }
-
-    uint16_t setMapSelectionSingleTile(const World::Pos2& loc, bool setQuadrant)
-    {
-        uint16_t xPos = loc.x & 0xFFE0;
-        uint16_t yPos = loc.y & 0xFFE0;
-        uint16_t cursorQuadrant = Ui::ViewportInteraction::getQuadrantOrCentreFromPos(loc);
-
-        auto count = 0;
-        if (!Input::hasMapSelectionFlag(Input::MapSelectionFlags::enable))
-        {
-            Input::setMapSelectionFlags(Input::MapSelectionFlags::enable);
-            count++;
-        }
-
-        if (setQuadrant && _word_F2448E != cursorQuadrant)
-        {
-            _word_F2448E = cursorQuadrant;
-            count++;
-        }
-        else if (!setQuadrant && _word_F2448E != 4)
-        {
-            _word_F2448E = 4;
-            count++;
-        }
-
-        if (xPos != _mapSelectionAX)
-        {
-            _mapSelectionAX = xPos;
-            count++;
-        }
-
-        if (yPos != _mapSelectionAY)
-        {
-            _mapSelectionAY = yPos;
-            count++;
-        }
-
-        if (xPos != _mapSelectionBX)
-        {
-            _mapSelectionBX = xPos;
-            count++;
-        }
-
-        if (yPos != _mapSelectionBY)
-        {
-            _mapSelectionBY = yPos;
-            count++;
-        }
-
-        mapInvalidateSelectionRect();
-
-        return count;
-    }
-
-    // 0x004610F2
-    void mapInvalidateSelectionRect()
-    {
-        if (Input::hasMapSelectionFlag(Input::MapSelectionFlags::enable))
-        {
-            for (coord_t x = _mapSelectionAX; x <= _mapSelectionBX; x += 32)
-            {
-                for (coord_t y = _mapSelectionAY; y <= _mapSelectionBY; y += 32)
-                {
-                    mapInvalidateTileFull({ x, y });
-                }
-            }
-        }
-    }
-
     // 0x004CBE5F
     // regs.ax: pos.x
     // regs.cx: pos.y
     void mapInvalidateTileFull(World::Pos2 pos)
     {
         Ui::ViewportManager::invalidate(pos, 0, 1120, ZoomLevel::eighth);
-    }
-
-    // 0x0046112C
-    void mapInvalidateMapSelectionTiles()
-    {
-        if (!Input::hasMapSelectionFlag(Input::MapSelectionFlags::enableConstruct))
-            return;
-
-        for (uint16_t index = 0; index < mapSelectedTilesSize; ++index)
-        {
-            auto& position = _mapSelectedTiles[index];
-            if (position.x == -1)
-                break;
-            mapInvalidateTileFull(position);
-        }
     }
 
     // 0x0046A747
